Handle error in nonblocking send
In a nonblocking send, we may be notified that the
connection was lost.  Currently this is lost because
the return value of flush_send_queue() is ignored.

This change DOES NOT change the ABI or the return
codes that are returned by zoo_a* async functions.

This change DOES close the file descriptor and
change zh->state to reflect that the connection was lost.